+++ ISAMRS_PROJEKAT/src/main/java/rest/domain/Farmaceut.java
@@ -9,7 +9,7 @@ public class Farmaceut extends Korisnik implements Ocenjivo{
 	
 	public Farmaceut() {}
 	
-	public Farmaceut(Long id, String ime, String prezime, String username, String password, String email,
+	public Farmaceut(int id, String ime, String prezime, String username, String password, String email,
 			Boolean loggedBefore, String telefon,ZaposlenjeKorisnika zaposlenjeKorisnika, Lokacija lokacija,
 			Zaposlenje zaposlenje,int brojOcena,int sumaOcena) {
 		super(id, ime, prezime, username, password, email, loggedBefore, telefon,lokacija,zaposlenjeKorisnika);
